Tests 2 with new dataset configuration for UNet.

+++ unet/lesionseg/data_augmentation/train_exp1_2.py
@@ -224,7 +224,7 @@ def train():
     imgs_train -= mean
     imgs_train /= std
 
-    imgs_mask_train = imgs_mask_train.astype(np.float32)
+    # imgs_mask_train = imgs_mask_train.astype(np.float32)
 
     # print('Train test split')
     # X_train, X_test, y_train, y_test = train_test_split(imgs_train, imgs_mask_train, test_size=0.1)
